removed useless float <-> double conversions

+++ pass/fplan/floorplanner.cpp
@@ -36,7 +36,7 @@ GeographyHint Lhd_floorplanner::randomHint(int count) const {
   return hint_seq[sel];
 }
 
-void Lhd_floorplanner::create(double ar) {
+void Lhd_floorplanner::create(float ar) {
   bool success = layouts[nt.root_index()]->layout(HardAspectRatio, ar);
   if (!success) {
     throw std::runtime_error("unable to lay out floorplan!");
